+++ packages/core/styles/src/packages/atomic/progressBar/index.ts
@@ -2,4 +2,4 @@ import * as styles from "./nimbus-progressBar.css";
 
 export const progressBar = {
   classnames: { ...styles },
-}; 
+};