+++ packages/core/typings/src/index.ts
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 import { ForwardedRef, RefObject, useImperativeHandle } from "react";
 
 /**