+++ packages/react/src/atomic/Button/src/button.types.ts
@@ -21,6 +21,11 @@ export interface ButtonProperties {
    * @default false
    */
   disabled?: boolean;
+  /**
+   * Determines if the button should grow to fill the full width of its container.
+   * @default false
+   */
+  fullWidth?: boolean;
 }
 
 export type ButtonBaseProps = ButtonProperties &