+++ packages/react/src/atomic/ProgressBar/src/index.ts
@@ -4,4 +4,4 @@ import { ProgressBar } from "./ProgressBar";
 
 export { ProgressBar } from "./ProgressBar";
 export type { ProgressBarProps } from "./ProgressBar";
-export default ProgressBar; 
+export default ProgressBar;